From Arvind I've attached the 4 source files with this, with updated code .. per our conversation in the concall earlier this week.
The code is supposed to support 4 new fields:
 MetricType (string, NULL OK)
 PerformanceData (string, NULL OK)
 VoName (string, NULL OK)
 SamUploadFlag (integer, NULL OK)

Note that the last field SamUploadFlag is supposed to be an integer field.
I used a data type called IntegerElement instead of StringElement which is
what the other fields (barring timeStamp) use. I don't know if I've done
this right, since I was kinda shooting in the dark with this one - please
verify my code mods and let me know if it looks appropriate.


git-svn-id: http://svn.code.sf.net/p/gratia/svn/trunk/probe@1266 c85a5f45-a5f3-44fc-b02c-7891c381c519