Pending files are now stored in a separate directory for each probes, still under the gratiafiles directory.
The directory is named subir.[ProbeName].[CollectorHost]
with the : and / replaced with _

The pending files are placed in a subdirectory of this probe specific
directory named 'outbox'.

Introduced the concept of staging of records:

When the outbox is full (i.e. number of files greater or equal to
MaxPendingFiles) instead of stopping the processing, we now
tar and bzip2 the content of outbox and place it in a subdirectory
named 'staged/stored'.   The archive is a simple bzip2 tar file
of the record without any pathname included.

When reprocessing file, we decompress each tar file (one at a time)
into the subdirectory 'staged/outbox'

The maximum number of tar file is limited by get_MaxStagedArchives.
When this maximum is reach, new record are rejected.

So the maximum number of physical files for each probes is
   2 * MaxPendingFiles + MaxStagedArchives

The maximum of outstanding records is
   2 * MaxPendingFiles + MaxStagedArchices*MaxPendingFiles

New Config Methods:
   FilenameFragment
   get_MaxStagedArchives

New statistics variables:
   OutstandingStagedRecordCount
   OutstandingStagedTarCount

New utilities routine:
   RemoveDir
   AddOutstandingRecord, ListOutstandingRecord, ReprocessList
   CompressOutbox, UncompressOutbox

New Bundle routine
   hasFile (to avoid adding a record already bundled)

Turn down the logging level 0


git-svn-id: http://svn.code.sf.net/p/gratia/svn/trunk/probe@3334 c85a5f45-a5f3-44fc-b02c-7891c381c519