Set the host description to the ProbeConfig SiteName for local jobs
In the case where a VO needs to account for locally run
jobs (i.e. SuppressGridLocalRecords="0"), we need to set the host
description to an OIM-registered name. In our documentation we tell
our users to set the SiteName in their ProbeConfig to the Resource
Name for their submit host so we use that.